+++ src/main/java/ru/isaykin/writer/MySQLWriter.java
@@ -39,7 +39,7 @@ public class MySQLWriter {
                 for (Author author : newTable) {
 
                     PreparedStatement preparedStatement1 = connection.prepareStatement(sqlReq);
-                    preparedStatement1.setInt(1, author.getId());
+                    preparedStatement1.setLong(1, author.getId());
                     preparedStatement1.setString(2, author.getFirstName());
                     preparedStatement1.setString(3, author.getLastName().replaceAll("'", "\\\\\'"));
                     preparedStatement1.setString(4, author.getEmail().replaceAll("'", "\\\\\'"));
